+++ tests/testthat/test-freq.R
@@ -25,3 +25,31 @@ test_that("frequency conversions return as expected", {
 
   expect_error(semitone_pitch(-1), "Semitones must range from 0 to 131.")
 })
+
+test_that("frequency ratios compile correctly", {
+  x <- as_music("c4 e_ g c'e_'g'")
+  fr <- freq_ratio(x)
+
+  x <- music_notes(x)
+  expect_identical(fr, freq_ratio(x))
+
+  x <- chord_freq(x)
+  fr1 <- freq_ratio(x)
+  expect_identical(fr, fr1)
+
+  fr2 <- freq_ratio(x, octaves = "integer", accidentals = "sharp")
+  expect_identical(fr1[, -2], fr2[, -2])
+  expect_equal(octave_type(fr2$notes), "integer")
+  expect_equal(accidental_type(fr2$notes), "sharp")
+  expect_identical(fr1$notes, as_noteworthy(fr2$notes, "tick", "flat"))
+
+  y <- freq_ratio(x, ratios = "root")
+  expect_identical(y$notes,
+                   as_noteworthy("c e_ g c'e_' c'g'", format = "vector"))
+  y <- freq_ratio(x, ratios = "range")
+  expect_identical(y$notes, as_noteworthy("c e_ g c'g'", format = "vector"))
+
+  x <- "c e_ g c'e_'g'"
+  fr3 <- freq_ratio(x)
+  expect_identical(fr, fr3)
+})