+++ NEWS.md
@@ -1,5 +1,6 @@
 # tabr 0.4.2
 
+* Added `freq_ratio` generic for generating a data frame of frequency ratios from frequencies, noteworthy objects, or music objects.
 * Added utility functions for retrieving LilyPond version and installation directory and `tabr` LilyPond API details.
 * Fixed bug where multiple sharps in tuning broke tab staff string label LilyPond syntax.
 * Updated documentation.